requirements.txt
requests==2.32.3
geopy==2.4.1
python-dotenv==1.0.1